+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Update University Housing Dining section name references in favorites [#5872](https://github.com/rokwire/illinois-app/issues/5872).
-- Updated MTD Bus Stops home panel [#5880](https://github.com/rokwire/illinois-app/issues/5880).
+- Update MTD Bus Stops home panel [#5880](https://github.com/rokwire/illinois-app/issues/5880).
+- Merge MTD Buses and Safety sections from Browse panel into new Campus Transit and Safety section [#5881](https://github.com/rokwire/illinois-app/issues/5881).
 
 ## [8.2.4] - 2026-08-14
 ### Fixed